src/requirements.txt
numpy==1.24.1
Pillow==9.4.0
pygame==2.1.2
python-dotenv==0.21.1
